Added DataList to the type system by default.
Moved array-determining deserialization logic out of the DataArray
class.

Fixed the return of the TypeSystem.createByCode.

Changed the TypeSystem container for types to an array.

Made sure the container can access negative bytes in the array, allowing
0x00 through 0xFF as all possible data codes.

If DataCompound.getString(name) was not a string return the toString
value of the held object.